refactor: extract modules from monolithic init.lua
Problem: `init.lua` had grown to ~1750 lines containing formatting,
template handling, config/types, and compose buffer logic all in one
file. This made the code hard to navigate and reason about.

Solution: Extract four focused modules — `format.lua` (formatting
helpers), `template.lua` (discovery/loading), `config.lua` (types,
defaults, validation), and `compose.lua` (PR/issue compose buffers) —
and reduce `init.lua` to a ~430-line orchestration layer that
re-exports submodule APIs for backward compatibility.
